+++ dubbo-common/src/main/java/org/apache/dubbo/common/json/JsonUtil.java
@@ -17,6 +17,7 @@
 package org.apache.dubbo.common.json;
 
 import org.apache.dubbo.common.extension.SPI;
+import org.apache.dubbo.config.ServiceConfigBase;
 
 import java.lang.reflect.Type;
 import java.util.List;
@@ -25,6 +26,8 @@ import java.util.Map;
 @SPI
 public interface JsonUtil {
     boolean isSupport();
+
+    boolean checkJsonCompatibility(ServiceConfigBase sc);
 
     <T> T toJavaObject(String json, Type type);
